# docs/swagger.yaml
openapi: 3.0.0
info:
  title: Multi-Project API with AI Integration
  description: |
    A multi-tenant API with AI-powered query translation and execution.
    This API allows users to manage projects, databases, and execute queries using natural language.
  version: 1.0.0
  contact:
    name: API Support
    email: support@example.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3000/api/v1
    description: Local development server
  - url: https://api.example.com/v1
    description: Production server
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >
        Enter the token with the `Bearer ` prefix, e.g. "Bearer abcde12345".
        
        You can obtain a token by authenticating with your credentials.
  schemas:
    Error:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: string
          description: Error message
          example: "Resource not found"
        timestamp:
          type: string
          format: date-time
          example: "2023-08-25T23:59:59.999Z"
...
    QueryResult:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: array
          items: {}
          description: Result rows from the query
        meta:
          type: object
          properties:
            execution_time_ms:
              type: number
              example: 15.25
            row_count:
              type: integer
              example: 5
security:
  - bearerAuth: []
paths:
...
  /query/ask:
    post:
      tags: [Queries]
      summary: Ask a question in natural language and get results
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - question
                - project_id
                - database_id
              properties:
                question:
                  type: string
                  example: "Show me the top 10 customers by total spending"
                project_id:
                  type: string
                  format: uuid
                  example: "550e8400-e29b-41d4-a716-446655440000"
                database_id:
                  type: string
                  format: uuid
                  example: "550e8400-e29b-41d4-a716-446655440000"
      responses:
        '200':
          description: Query executed successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/QueryResult'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /query/translate:
    post:
      tags: [Queries]
      summary: Translate natural language to database query
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - query
                - project_id
              properties:
                query:
                  type: string
                project_id:
                  type: string
                  format: uuid
                database_id:
                  type: string
                  format: uuid
                schema:
                  type: object
      responses:
        '200':
          description: Query translated successfully
        '400':
          description: Invalid request
  /query/execute:
    post:
      tags: [Queries]
      summary: Execute a database query
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - query
                - project_id
                - database_id
              properties:
                query:
                  type: string
                project_id:
                  type: string
                  format: uuid
                database_id:
                  type: string
                  format: uuid
                params:
                  type: array
                is_ai_generated:
                  type: boolean
                  default: false
      responses:
        '200':
          description: Query executed successfully
        '400':
          description: Invalid request or query
        '403':
          description: Insufficient permissions
  /query/history:
    get:
      tags: [Queries]
      summary: Get query history for a project
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: project_id
          required: true
          schema:
            type: string
            format: uuid
        - in: query
          name: limit
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - in: query
          name: offset
          schema:
            type: integer
            minimum: 0
            default: 0
      responses:
        '200':
          description: Query history retrieved successfully
        '400':
          description: Invalid request